final commit of cmsc 350
The code works as I want. The constructor sorts the polynomial as the file is built. The to String displays the correct syntax in the console. A little more work would need to be done in order to finish the code for submission, such as implementing the compareTo method. However, the compareTo method is finished and the code works.